+++ packages/iov-core/src/index.ts
@@ -25,5 +25,7 @@ export {
   parseJsRpcResponse,
 } from "./jsrpc";
 export { JsRpcSigningServer } from "./jsrpcsigningserver";
+export { JsonRpcSigningServer } from "./jsonrpcsigningserver";
 export { MultiChainSigner } from "./multichainsigner";
 export { GetIdentitiesAuthorization, SignAndPostAuthorization, SigningServerCore } from "./signingservercore";
+export { TransactionEncoder } from "./transactionencoder";