+++ packages/iov-jsonrpc/types/types.d.ts
@@ -25,9 +25,30 @@ export interface JsonRpcErrorResponse {
 }
 export declare type JsonRpcResponse = JsonRpcSuccessResponse | JsonRpcErrorResponse;
 export declare function isJsonRpcErrorResponse(response: JsonRpcResponse): response is JsonRpcErrorResponse;
-export declare const jsonRpcCodeParseError = -32700;
-export declare const jsonRpcCodeInvalidRequest = -32600;
-export declare const jsonRpcCodeMethodNotFound = -32601;
-export declare const jsonRpcCodeInvalidParams = -32602;
-export declare const jsonRpcCodeInternalError = -32603;
-export declare const jsonRpcCodeServerErrorDefault = -32000;
+/**
+ * Error codes as specified in JSON-RPC 2.0
+ *
+ * @see https://www.jsonrpc.org/specification#error_object
+ */
+export declare const jsonRpcCode: {
+    parseError: number;
+    invalidRequest: number;
+    methodNotFound: number;
+    invalidParams: number;
+    internalError: number;
+    serverError: {
+        default: number;
+    };
+};
+/** @deprecated: use jsonRpcCode */
+export declare const jsonRpcCodeParseError: number;
+/** @deprecated: use jsonRpcCode */
+export declare const jsonRpcCodeInvalidRequest: number;
+/** @deprecated: use jsonRpcCode */
+export declare const jsonRpcCodeMethodNotFound: number;
+/** @deprecated: use jsonRpcCode */
+export declare const jsonRpcCodeInvalidParams: number;
+/** @deprecated: use jsonRpcCode */
+export declare const jsonRpcCodeInternalError: number;
+/** @deprecated: use jsonRpcCode */
+export declare const jsonRpcCodeServerErrorDefault: number;