+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## 0.13.4
 
 - @iov/keycontrol: Export `ReadonlyWallet`.
+- @iov/jsonrpc: Add `jsonRpcCode` constant object. Deprecate old `jsonRpcCode*`
+  constants in favour of `jsonRpcCode`.
+- @iov/core: Add `JsonRpcSigningServer`.
+- @iov/core: Deprecate `JsRpcSigningServer` and related types in favour of
+  `JsonRpcSigningServer`.
 
 ## 0.13.3